+++ packages/core/styles/src/packages/atomic/box/box.sprinkle.css.ts
@@ -18,6 +18,9 @@ import {
   flexWrapProperties,
   alignItemsProperties,
   pointerEventsProperties,
+  boxShadowProperties,
+  transitionSpeedProperties,
+  transitionTimingProperties,
 } from "../../../properties";
 import { varsThemeBase, mediaQueries } from "../../../themes";
 
@@ -51,7 +54,10 @@ const boxBackgroundColorProperties = {
   "neutral.surface": backgroundColorProperties["neutral.surface"],
   "neutral.surfaceHighlight":
     backgroundColorProperties["neutral.surfaceHighlight"],
+  "neutral.surfaceDisabled":
+    backgroundColorProperties["neutral.surfaceDisabled"],
   "neutral.textLow": backgroundColorProperties["neutral.textLow"],
+  transparent: "transparent",
 };
 
 const boxBorderColorProperties = {
@@ -74,6 +80,7 @@ const boxBorderColorProperties = {
   "neutral.interactive": borderColorProperties["neutral.interactive"],
   "neutral.surface": borderColorProperties["neutral.surface"],
   "neutral.surfaceHighlight": borderColorProperties["neutral.surfaceHighlight"],
+  transparent: "transparent",
 };
 
 const boxSpaceProperties = {
@@ -112,6 +119,7 @@ const defineProperties = defineRainbowProperties({
     active: { selector: "&:active" },
     hover: { selector: "&:hover" },
     focus: { selector: "&:focus" },
+    disabled: { selector: "&:disabled" },
   },
   defaultCondition: "xs",
   dynamicProperties: {
@@ -131,8 +139,12 @@ const defineProperties = defineRainbowProperties({
     flexShrink: true,
     borderRadius: true,
     borderWidth: true,
+    transitionDelay: true,
+    transitionProperty: true,
+    transitionDuration: transitionSpeedProperties,
     backgroundColor: boxBackgroundColorProperties,
     borderColor: boxBorderColorProperties,
+    boxShadow: boxShadowProperties,
     padding: boxSpaceProperties,
     paddingTop: boxSpaceProperties,
     paddingBottom: boxSpaceProperties,
@@ -160,6 +172,7 @@ const defineProperties = defineRainbowProperties({
     overflow: overflowProperties,
     overflowX: overflowProperties,
     overflowY: overflowProperties,
+    transitionTimingFunction: transitionTimingProperties,
   },
   shorthands: {
     padding: ["paddingTop", "paddingBottom", "paddingLeft", "paddingRight"],
@@ -205,6 +218,9 @@ const properties = {
   overflow: overflowProperties,
   overflowX: overflowProperties,
   overflowY: overflowProperties,
+  boxShadow: boxShadowProperties,
+  transitionDuration: transitionSpeedProperties,
+  transitionTimingFunction: transitionTimingProperties,
 };
 
 const sprinkle = createRainbowSprinkles(defineProperties);